refactor: Update import paths for Donut chart components
