+++ frontend/src/components/UI/DataViz/ResponsiveDonutWithInnerPercent/CustomLayerComponent.jsx
